Implemented Menu and NavBar of Dashboard
I separated out Dashboard to different components. Create.js is the button to create a new project, Dashboard.js is the main file for dashboard. Using material UI to implement.
