+++ package.json
@@ -7,9 +7,13 @@
     "@babel/core": "^7.4.4",
     "@babel/preset-env": "^7.4.4",
     "@babel/preset-react": "^7.0.0",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
     "@hot-loader/react-dom": "^16.8.6",
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/material": "^5.8.6",
     "@octokit/auth-app": "^3.4.0",
     "@octokit/rest": "^18.5.3",
     "all": "^0.0.0",
